+++ CHANGELOG.md
@@ -8,6 +8,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - clean async process for import analysis repository
 - edit user able to upload file like CV, diploma ... as for edit staff GUI
 
+## [3.6.1] - 2025-09-29
+### Added
+- new dictionary for equipment status
+- equipment status in details, list and export of equipment
+- filter for include, exclude or only LabBook Lite records for DHIS2 export
+- show Biological Validation button on Enter Results and Technical Validation pages if the user has permission and at least one result is technically validated
+- Modify patient button in administrative record page
+
 ## [3.6.0] - 2025-09-24
 ### Added
 - new setting to indicate whether a variable should be displayed in the report